requirements.txt
beautifulsoup4
Flask
requests
mysql-connector-python
python-dotenv


-e .